src/tools/customModel/requirements.txt
requests
azure-storage-blob
azure-core